EditModeBar created here Added ImageLoader with disk caching enabled to load images fro StaffAvatarItem
Implemented click action for StaffDetailCard to show menu dialog which launches StaffListItemMenuDialog

Added StaffListDetailRow for StaffListScreen

changed the scrollIndex of MonthListColumn with (-2) to make sure it's not scrolled to the top of the screen.

Fixed the progress logic for the progressbar, and added an icon to show, when in edit mode

Added:
CallTypeSegmentedButton for edit mode in the CallRosterScreen
MainResourceScreen for the resources screen (with ResourceRow)
DetailScreen for the resources screen sub screen (with OrganogramScreen, ClerkingScreen, GuideSection, SystemicCG, SegmentalCG, PediatricsCG,
NotificationsScreen, showSettingsDialog (to prompt opening settings to enable notifications on phone), NotificationsStaffItem, ReminderItem, SuggestionsScreen, SuggestionItem, AboutScreen and ClickableTextItem all included.